Refactor drag and drop into hook
