+++ lightly_studio_view/src/lib/components/GridItem/GridItem.svelte
@@ -1,11 +1,8 @@
 <script lang="ts">
-    import { onDestroy } from 'svelte';
     import type { Snippet } from 'svelte';
     import GridItemTag from './GridItemTag.svelte';
-    import { GRID_IMAGE_SEARCH_DROP_EVENT, type GridItemDragData } from './GridItem.constants';
-    const GRID_IMAGE_SEARCH_DROP_TARGET_SELECTOR = '[data-grid-search-drop-target]';
-    const DRAG_START_THRESHOLD_PX = 8;
-    const DRAG_PREVIEW_OFFSET_PX = 14;
+    import { DRAG_PREVIEW_OFFSET_PX, type GridItemDragData } from './GridItem.constants';
+    import { useGridItemDrag } from '$lib/hooks/useGridItemDrag/useGridItemDrag.svelte';
 
     let {
         children,
@@ -39,11 +36,7 @@
         ondblclick?: (event: MouseEvent) => void;
     } = $props();
 
-    let suppressNextClick = false;
-    let pointerStart: { x: number; y: number; id: number } | null = null;
-    let capturedPointerElement: HTMLElement | null = null;
-    let isPointerDragging = $state(false);
-    let dragPreview = $state<{ x: number; y: number } | null>(null);
+    const drag = useGridItemDrag(() => dragData);
 
     function formatSize(value: string | number): string {
         return typeof value === 'number' ? `${value}px` : value;
@@ -51,8 +44,8 @@
 
     function handleOnClick(event: MouseEvent) {
         if (!onSelect) return;
-        if (suppressNextClick) {
-            suppressNextClick = false;
+        if (drag.suppressNextClick) {
+            drag.suppressNextClick = false;
             event.preventDefault();
             return;
         }
@@ -66,121 +59,6 @@
         event.preventDefault();
         onSelect(event);
     }
-
-    function suppressNextSelectionClick() {
-        suppressNextClick = true;
-        setTimeout(() => {
-            suppressNextClick = false;
-        }, 100);
-    }
-
-    function getPointerDistance(event: PointerEvent): number {
-        if (!pointerStart) {
-            return 0;
-        }
-        return Math.hypot(event.clientX - pointerStart.x, event.clientY - pointerStart.y);
-    }
-
-    function getPointerTarget(event: PointerEvent): HTMLElement | null {
-        return event.currentTarget instanceof HTMLElement ? event.currentTarget : null;
-    }
-
-    function getSearchDropTarget(event: PointerEvent): Element | null {
-        return (
-            document
-                .elementFromPoint(event.clientX, event.clientY)
-                ?.closest(GRID_IMAGE_SEARCH_DROP_TARGET_SELECTOR) ?? null
-        );
-    }
-
-    function setBodyDraggingCursor(isDragging: boolean) {
-        document.body.style.cursor = isDragging ? 'grabbing' : '';
-    }
-
-    function updateDragPreview(event: PointerEvent) {
-        dragPreview = { x: event.clientX, y: event.clientY };
-    }
-
-    function releaseCapturedPointer(pointerId = pointerStart?.id) {
-        if (capturedPointerElement && pointerId != null) {
-            try {
-                if (capturedPointerElement.hasPointerCapture?.(pointerId)) {
-                    capturedPointerElement.releasePointerCapture(pointerId);
-                }
-            } catch {
-                // Pointer capture may already be gone when cleanup races browser events.
-            }
-        }
-        capturedPointerElement = null;
-    }
-
-    function stopDragging() {
-        releaseCapturedPointer();
-        pointerStart = null;
-        isPointerDragging = false;
-        dragPreview = null;
-        setBodyDraggingCursor(false);
-    }
-
-    function handlePointerDown(event: PointerEvent) {
-        if (!dragData || event.button !== 0) {
-            return;
-        }
-        pointerStart = { x: event.clientX, y: event.clientY, id: event.pointerId };
-        isPointerDragging = false;
-        capturedPointerElement = getPointerTarget(event);
-        capturedPointerElement?.setPointerCapture?.(event.pointerId);
-    }
-
-    function handlePointerMove(event: PointerEvent) {
-        if (!dragData || !pointerStart || pointerStart.id !== event.pointerId) {
-            return;
-        }
-        if (!isPointerDragging && getPointerDistance(event) < DRAG_START_THRESHOLD_PX) {
-            return;
-        }
-        if (!isPointerDragging) {
-            isPointerDragging = true;
-            suppressNextSelectionClick();
-            setBodyDraggingCursor(true);
-        }
-        updateDragPreview(event);
-    }
-
-    function handlePointerUp(event: PointerEvent) {
-        if (!dragData || !pointerStart || pointerStart.id !== event.pointerId) {
-            stopDragging();
-            return;
-        }
-
-        const dropTarget = isPointerDragging ? getSearchDropTarget(event) : null;
-
-        if (isPointerDragging) {
-            suppressNextSelectionClick();
-        }
-
-        stopDragging();
-
-        if (dropTarget) {
-            window.dispatchEvent(
-                new CustomEvent<GridItemDragData>(GRID_IMAGE_SEARCH_DROP_EVENT, {
-                    detail: dragData
-                })
-            );
-        }
-    }
-
-    function handlePointerCancel() {
-        stopDragging();
-    }
-
-    function handleLostPointerCapture() {
-        stopDragging();
-    }
-
-    onDestroy(() => {
-        stopDragging();
-    });
 </script>
 
 <div class="select-none" {style}>
@@ -192,15 +70,15 @@
         data-sample-name={dataSampleName}
         data-index={dataIndex}
         draggable="false"
-        class:cursor-grab={dragData && !isPointerDragging}
-        class:cursor-grabbing={isPointerDragging}
+        class:cursor-grab={dragData && !drag.isPointerDragging}
+        class:cursor-grabbing={drag.isPointerDragging}
         {ondblclick}
         onclick={handleOnClick}
-        onpointerdown={handlePointerDown}
-        onpointermove={handlePointerMove}
-        onpointerup={handlePointerUp}
-        onpointercancel={handlePointerCancel}
-        onlostpointercapture={handleLostPointerCapture}
+        onpointerdown={drag.handlePointerDown}
+        onpointermove={drag.handlePointerMove}
+        onpointerup={drag.handlePointerUp}
+        onpointercancel={drag.handlePointerCancel}
+        onlostpointercapture={drag.handleLostPointerCapture}
         onkeydown={handleKeyDown}
         aria-label={ariaLabel}
         role="button"
@@ -225,10 +103,10 @@
     </div>
 </div>
 
-{#if dragData && dragPreview}
+{#if dragData && drag.dragPreview}
     <div
         class="pointer-events-none fixed z-50 h-20 w-20 overflow-hidden rounded-md border border-primary/60 bg-background opacity-80 shadow-xl ring-2 ring-primary/30"
-        style="left: {dragPreview.x + DRAG_PREVIEW_OFFSET_PX}px; top: {dragPreview.y +
+        style="left: {drag.dragPreview.x + DRAG_PREVIEW_OFFSET_PX}px; top: {drag.dragPreview.y +
             DRAG_PREVIEW_OFFSET_PX}px;"
         data-testid="grid-item-drag-preview"
     >
